Krivdin Mikhail b57

+++ src/task3.cpp
@@ -3,7 +3,7 @@
 
 unsigned long long sumPrime(unsigned int hbound)
 {
-	int sum = 0;
+	unsigned long long sum = 0;
 	for (unsigned long long i = 2; i < hbound; i++)
 	{
 		if (checkPrime(i))
